migrate test to wallet.rs
